+++ packages/core-js/es/error/constructor.js
@@ -1,0 +1,4 @@
+require('../../modules/es.error.cause');
+var path = require('../../internals/path');
+
+module.exports = path;