+++ packages/core-js/proposals/error-cause.js
@@ -1,0 +1,3 @@
+// https://github.com/tc39/proposal-error-cause
+require('../modules/es.error.cause');
+require('../modules/es.aggregate-error.cause');