+++ packages/core-js/stable/error/constructor.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/error/constructor');
+
+module.exports = parent;