+++ packages/core-js/stable/error/index.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/error');
+
+module.exports = parent;