+++ packages/core-js/stage/4.js
@@ -1,4 +1,5 @@
 require('../proposals/accessible-object-hasownproperty');
+// require('../proposals/error-cause');
 require('../proposals/global-this');
 require('../proposals/promise-all-settled');
 require('../proposals/promise-any');