+++ tests/commonjs.mjs
@@ -119,6 +119,8 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok('map' in load(NS, 'array/virtual'));
     ok('from' in load(NS, 'array'));
     ok(load(NS, 'array/splice')([1, 2, 3], 1, 2)[0] === 2);
+    ok(load(NS, 'error/constructor').Error(1, { cause: 7 }).cause === 7);
+    ok(load(NS, 'error').Error(1, { cause: 7 }).cause === 7);
     ok(load(NS, 'math/acosh')(1) === 0);
     ok(Object.is(load(NS, 'math/asinh')(-0), -0));
     ok(load(NS, 'math/atanh')(1) === Infinity);
@@ -799,6 +801,7 @@ for (PATH of ['core-js-pure', 'core-js']) {
   load('proposals/collection-of-from');
   load('proposals/decorators');
   load('proposals/efficient-64-bit-arithmetic');
+  load('proposals/error-cause');
   load('proposals/function-is-callable-is-constructor');
   load('proposals/function-un-this');
   load('proposals/global-this');